+++ client/src/pages/Login.js
@@ -1,0 +1,12 @@
+import React from "react";
+import Frame from "../components/Frame";
+
+function Login() {
+  return (
+    <div>
+      <Frame h1="Login" h2="Sign in to continue" />
+    </div>
+  );
+}
+
+export default Login;