+++ CHANGELOG.md
@@ -5,12 +5,33 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.3] - 2022-06-15
+### Added
+- Update, Error, and Login pages.
+- Button and BasicForm components.
+
+### Modified
+- Home page deconstructed into Frame component.
+
+## [0.0.2] - 2022-06-14
+### Added
+- router and transition dependencies for client.
+- parser dependency for server.
+- home page, button component for client.
+- css modules for home page and button component.
+- favicon.
+
+### Modified
+- slight changes to package.json in client and server.
+- path directories in manifest.json and index.html for client.
+
 ## [0.0.1] - 2022-06-14
 ### Project created.
 
 ### Added
-- index, admin, and login html files.
-- style.css file.
-- image and js folders.
+- client & server folders
+- installed initial dependencies
+- setup readme, license, and changelog files
+- base server.js file
 
 [0.0.1]: https://github.com/